refactor: extract draft management to useDraftManager hook
- Moved 126 lines of draft logic from shorts.tsx to hooks/useDraftManager.ts
- Reduced shorts.tsx from 642 to 350 lines (45% smaller)
- Clean separation: UI vs business logic
- Improved maintainability and testability
